lmdb: split msgfunc related stuff into msgfunc.go
The previous placement in lmdb.go was somewhat of an accident.  And the
separation provides a clear distinction for C dependencies.
